changes to sign-up & log-in, adds a check to create account page to ensure all fields have been filled in before trying to create an account, changes username check to on blur to give better feedback to users

+++ src/components/Invitations/CreatedInvitations.tsx
@@ -84,7 +84,7 @@ export const CreatedInvitations: React.FC<InvitationsProps> = ({
                     <button
                       className="btn btn-primary"
                       onClick={() => {
-                        setInviteToDelete(invite.id)
+                        setInviteToDelete(invite.id);
                         const modal = document.getElementById(
                           "my_modal_1"
                         ) as HTMLDialogElement;
@@ -107,7 +107,9 @@ export const CreatedInvitations: React.FC<InvitationsProps> = ({
                           <button
                             className="btn btn-outline btn-primary mr-5"
                             onClick={() => {
-                              handleDelete(inviteToDelete);
+                              if (inviteToDelete) {
+                                handleDelete(inviteToDelete);
+                              }
                             }}
                           >
                             Confirm
